apps/ai-agent-example/requirements.txt
websocket-client>=1.6.0
requests>=2.31.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

